+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - Domain
     - Type
     - Cases
+- Add `KeyWord` `Do` and parse it as a `Parsed.KeyWordOnly`
 
 ## 2.0.1 - 2020-11-06
 - Add missing `KeyWord` value to `KeyWordIf`